+++ src/ScottPlot5/ScottPlot5/Primitives/FontStyle.cs
@@ -1,6 +1,4 @@
-﻿using System.Runtime.InteropServices.ComTypes;
-
-namespace ScottPlot;
+﻿namespace ScottPlot;
 
 /// <summary>
 /// This configuration object (reference type) permanently lives inside objects which require styling.
@@ -8,60 +6,11 @@ namespace ScottPlot;
 /// </summary>
 public class FontStyle
 {
-    /* Typefaces are cached to improve performance.
-     * https://github.com/ScottPlot/ScottPlot/issues/2833
-     * https://github.com/ScottPlot/ScottPlot/pull/2848
-     */
+    public SKTypeface Typeface => Fonts.GetTypeface(Name, Bold, Italic);
 
-    private SKTypeface? CachedTypeface = null;
-
-    // TODO: use a class for cached typeface management
-    public SKTypeface Typeface => CachedTypeface ??= CreateTypefaceFromName(Name, Bold, Italic);
-
-    private string _name = Fonts.Default;
-    public string Name
-    {
-        get => _name;
-        set
-        {
-            bool fieldChanged = string.Compare(_name, value, StringComparison.InvariantCultureIgnoreCase) != 0;
-
-            if (fieldChanged)
-                ClearCachedTypeface();
-
-            _name = value;
-        }
-    }
-
-    private bool _bold = false;
-    public bool Bold
-    {
-        get => _bold;
-        set
-        {
-            bool fieldChanged = _bold != value;
-
-            if (fieldChanged)
-                ClearCachedTypeface();
-
-            _bold = value;
-        }
-    }
-
-    private bool _italic = false;
-    public bool Italic
-    {
-        get => _italic;
-        set
-        {
-            bool fieldChanged = (_italic != value);
-
-            if (fieldChanged)
-                ClearCachedTypeface();
-
-            _italic = value;
-        }
-    }
+    public string Name { get; set; } = Fonts.Default;
+    public bool Bold { get; set; } = false;
+    public bool Italic { get; set; } = false;
 
     // TODO: consider whether color really belongs here...
     public Color Color { get; set; } = Colors.Black;
@@ -73,26 +22,16 @@ public class FontStyle
         return $"{Name}, Size {Size}, {Color}";
     }
 
-    private void ClearCachedTypeface()
-    {
-        CachedTypeface = null;
-    }
-
+    [Obsolete("This method is deprecated. Use Fonts.GetTypeface(font, bold, italic) instead.", true)]
     public static SKTypeface CreateTypefaceFromName(string font, bool bold, bool italic)
     {
-        SKFontStyleWeight weight = bold ? SKFontStyleWeight.Bold : SKFontStyleWeight.Normal;
-        SKFontStyleSlant slant = italic ? SKFontStyleSlant.Italic : SKFontStyleSlant.Upright;
-        SKFontStyleWidth width = SKFontStyleWidth.Normal;
-        SKFontStyle style = new(weight, width, slant);
-        return SKTypeface.FromFamilyName(font, style);
+        throw new NotImplementedException();
     }
 
+    [Obsolete("This method is deprecated. Use Fonts.GetTypeface(font, bold, italic) instead.", true)]
     public static SKTypeface CreateTypefaceFromFile(string path)
     {
-        path = Path.GetFullPath(path);
-        if (!File.Exists(path))
-            throw new FileNotFoundException(path);
-        return SKTypeface.FromFile(path);
+        throw new NotImplementedException();
     }
 
     /// <summary>