+++ src/ScottPlot5/ScottPlot5/Primitives/Label.cs
@@ -32,55 +32,9 @@ public class Label
     public bool AntiAliasBackground { get; set; } = true;
     public bool AntiAliasText { get; set; } = true;
 
-    // TODO: use a class for cached typeface management
-
-    private SKTypeface? CachedTypeface = null;
-    private SKTypeface Typeface
-    {
-        get
-        {
-            if (CachedTypeface is not null)
-                return CachedTypeface;
-
-            if (FontFile is not null)
-                return FontStyle.CreateTypefaceFromFile(FontFile);
-
-            return FontStyle.CreateTypefaceFromName(FontName, Bold, Italic);
-        }
-    }
-
-    public void SetTypeface(SKTypeface typeface)
-    {
-        CachedTypeface = typeface;
-    }
-
-    private string _FontName = Fonts.Default;
-    public string FontName
-    {
-        get => _FontName;
-        set { _FontName = value; ClearCachedTypeface(); }
-    }
-
-    private string? _FontFile = null;
-    public string? FontFile
-    {
-        get => _FontFile;
-        set { _FontFile = value; ClearCachedTypeface(); }
-    }
-
-    private float _FontSize = 12;
-    public float FontSize
-    {
-        get => _FontSize;
-        set { _FontSize = value; ClearCachedTypeface(); }
-    }
-
-    private bool _Bold = false;
-    public bool Bold
-    {
-        get => _Bold;
-        set { _Bold = value; ClearCachedTypeface(); }
-    }
+    public string FontName { get; set; } = Fonts.Default;
+    public float FontSize { get; set; } = 12;
+    public bool Bold { get; set; } = false;
 
     /// <summary>
     /// Manually defined line height in pixels.
@@ -123,10 +77,6 @@ public class Label
         FontName = Fonts.Detect(Text);
     }
 
-    public void ClearCachedTypeface()
-    {
-        CachedTypeface = null;
-    }
 
     private void ApplyPointPaint(SKPaint paint)
     {
@@ -166,7 +116,7 @@ public class Label
     {
         paint.TextAlign = SKTextAlign.Left;
         paint.IsStroke = false;
-        paint.Typeface = Typeface;
+        paint.Typeface = Fonts.GetTypeface(FontName, Bold, Italic);
         paint.TextSize = FontSize;
         paint.Color = ForeColor.ToSKColor();
         paint.IsAntialias = AntiAliasText;